Unit test for Student Epic #77
