+++ js/User/luhnCheck.js
@@ -1,0 +1,16 @@
+function luhnCheck(num) {
+  if (!/^\d{13,19}$/.test(num)) return false;
+  let sum = 0;
+  let alt = false;
+  for (let i = num.length - 1; i >= 0; i--) {
+    let n = parseInt(num.charAt(i), 10);
+    if (alt) {
+      n *= 2;
+      if (n > 9) n -= 9;
+    }
+    sum += n;
+    alt = !alt;
+  }
+  return sum % 10 === 0;
+}
+module.exports = luhnCheck;
